+++ lib/torrent_parser/__init__.py
@@ -1,0 +1,1 @@
+from .torrent_parser import *